pyproject.toml
[tool.poetry]
name = "back-end"
version = "0.1.0"
description = ""
authors = ["Allen Anthes <allen@operationcode.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
bcrypt = "^3.1"
boto3 = "^1.9"
django = "^2.2"
djangorestframework = "3.9"
djangorestframework-camel-case = "^1.0"
djangorestframework-jwt = "^1.11"
django-allauth = "^0.39.1"
django-anymail = "^6.1"
django-background-tasks = "^1.2"
django-cors-headers = "^3.0"
django-recaptcha2 = "^1.4"
django-rest-auth = "^0.9.5"
django-split-settings = "^0.3.0"
django-storages = "^1.7"
django-suit = "^0.2.26"
django-suit-daterange-filter = "^0.0.6"
django-widget-tweaks = "^1.4"
drf-yasg = "^1.16"
gunicorn = "^19.9"
packaging = "^19.1"
pillow = "^6.0"
psycopg2 = "^2.8"
python-decouple = "^3.1"
mailchimp3 = "^3.0"
mandrill = "^1.0"
python-dotenv = "^0.10.3"
requests = "^2.21"
sentry-sdk = "^0.10.1"
six = "^1.12"

[tool.poetry.dev-dependencies]
bandit = "^1.6"
black = {version = "^19.3b0", allows-prereleases = true}
coverage = "^4.5"
django-debug-toolbar = "^2.0"
factory_boy = "^2.12"
flake8 = "^3.7"
flake8-bugbear = "^19.3"
isort = "^4.3"
pyhumps = "^1.2"
pytest = "^5.0"
pytest-django = "^3.5"
pytest-env = "^0.6.2"
pytest-mock = "^1.10"
pytest-pythonpath = "^0.7.3"
responses = "^0.10.6"

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

